created navbar component
